pyproject.toml
[tool.poetry]
name = "sportbot"
version = "0.1.0"
description = "SportBox personal trainer"
authors = ["Adrien Pensart <adrien.pensart@corp.ovh.com>"]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
gTTS = "^2.2.2"
python-slugify = "^8.0.0"
progressbar2 = "^4.4.2"
python-cfonts = "^1.5.2"
click-skeleton = "^0.30"
pendulum = "^3.0.0"
prompt-toolkit = "^3.0.20"
colorlog = "^6.4.1"
beartype = "^0.18.5"
pyobjc = {version = "^7.3", markers = "sys_platform == 'darwin'"}
PyGObject = {version = "^3.40.1", markers = "sys_platform == 'linux'"}

[tool.poetry.group.dev.dependencies]
pylint = "^3.2.2"
mypy = "^1.10.0"
types-python-slugify = "^8.0.0"
restructuredtext-lint = "^1.3.2"
pytest = "^8.2.1"
pytest-cov = "^5.0.0"
black = "^24.0.0"
isort = "^5.10.1"
ruff = "<1.0"
flake8 = "^7.0.0"

[tool.isort]
profile = 'black'

[tool.black]
line-length = 180

[tool.ruff]
lint.ignore = ["E501"]

[tool.poetry.scripts]
sportbot = 'sportbot.main:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--cov-report term-missing --durations=0 --cov sportbot "

[tool.pylint.master]
jobs = 4
[tool.pylint.messages_control]
disable = """
too-few-public-methods,logging-format-interpolation,line-too-long,too-many-arguments,protected-access,missing-docstring,invalid-name,too-many-public-methods,too-many-instance-attributes,duplicate-code,too-many-nested-blocks,too-many-branches,too-many-return-statements,too-many-statements,too-many-locals,too-few-public-methods,too-many-ancestors,abstract-method,anomalous-backslash-in-string,import-outside-toplevel,redefined-outer-name,unnecessary-lambda,c-extension-no-member,logging-fstring-interpolation
"""

[tool.mypy]
allow_redefinition = false
disallow_untyped_defs = true
ignore_errors = false
implicit_reexport = false
local_partial_types = true
strict_optional = true
strict_equality = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unreachable = true
warn_no_return = true